Add support for controlling OpenSSL's FIPS 140-2 module
Adds the tls::FIPS_mode and tls::FIPS_mode_set commands to the module
which query and set the operation of OpenSSL's FIPS 140-2 module
respectively.  These are conditionally compiled in on the OPENSSL_FIPS
macro which is enabled through a new configure flag,
--enable-fips-control

The commands simply expose the underlying FIPS_mode and FIPS_mode_set
C routines from the OpenSSL library.